Start reorganising Action/Menu/TitleBar factory to be more generic and more capable of use within an IoC environment.
At this time the ActionFactory is the focus of attention.

The old ActionFactory became DelegationActionFactory.
A new ActionFactory interface was created in it's place.

DelegationActionFactory, via AbstractActionFactory now uses a ResourceLocator (see stock-common-iface 1.3) to obtain icons.  The ImageCache class now extends IconLocator, although it's support for ResourceLocator methods is unimplemented, and can be used by an ActionFactory.java

The default behaviour for AbstractActionFactory is to use a ClassResourceLocator, which differs from the previous incarnation which used a SingletonImageCache.