Show ALL errors. This was odd - the count of errors was counting all errors, but only displaying the "base" errors, not the per-attribute ones.
I fixed it to show the errors in accordance with the count.
Perhaps it should have been fixed by only reporting the number of "base" errors.  I didn't think this was a good idea because then you might have no apparent errors when there are no base and some per-attribute errors.