Fix failures in integration-testing of acceptance tests relating to management authorities.
The default for mapping of urls when using the "ControllerClassNameHandlerMapping", as we are, requires the url to contain the lowercase version of the controller name - thus "managementauthoritys".
Unfortunately the name of the JSP that was handling the output of this url was called "managementAuthoritys.jsp".
The problem is that jetty is different to tomcat when it comes to case-sensitivity in filenames and testing under jetty was passing but not under Tomcat.
Tomcat expects a filename that matches the url - thus a camel-case file won't match the lowercase url.
Jetty appears to compile all JSP files to classes with an entirely lowercased name - which thus matches the lowercase url.
Specifying the camel-case name of the JSP in the controller fixes Tomcat and breaks Jetty.

Reworked ManagementAuthorityController to be a multi action controller.

This has the side effect of fixing the failure of the acceptance test by virtue of making the view page be managementauthority/list.jsp

In future the case filenames need to be very carefully checked - and pretty much always set to lowercase.