The recent update to use spring 2.5.3 has invalidated a large number of the jars held in the 'lib' dir. So I've completely changed my mind about how we handle lib files for this project.
- Remove and ignore the lib directory
- Add configuration for the maven-dependency-plugin to the pom
- Rework the IDEA module file to look for all libs in a dir called 'lib', without all the subdirectories.

People using IDEA should create the contents of the lib dir using the maven command:
   mvn dependency:copy-dependencies
This will copy all dependency jars from their personal repository into the lib dir, where the IDEA Module file will happily find them.

Come the end of the project when we do a final release we might checking the contents of the lib dir as a long-term security thing, or we can configure a maven task which will include all the dependencies in a source distribution package (this seems like a better option).