Remove the use of hibernate for management of the database schema and dbunit for database population, replacing them with the use of the active record migrations.
Remove the need for a hand-built config/database.yml, although this is still supported.
A database configuration file will be built using the example.database.yml and the values from the pom.xml as part of the maven build.

Sort of add support in the pom at least for a separate test database compared to the development database, but it's not properly handled by appfuse so this isn't having any real effect.